% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backSample.R
\name{backSample}
\alias{backSample}
\title{backSample}
\usage{
backSample(x, y, z, sampling.method = "random", nr.samples = NULL)
}
\arguments{
\item{x}{Object of class \emph{SpatialPoints} of \emph{SpatialPointsDataFrame}.}

\item{y}{Object of class \emph{RasterLayer}, \emph{RasterStack} or \emph{RasterBrick}.}

\item{z}{Vector of region identifiers for each sample.}

\item{sampling.method}{One of \emph{random} or \emph{pca}. Default is \emph{random}.}

\item{nr.samples}{Number of random background samples.}
}
\value{
A \emph{SpatialPoints} or a \emph{SpatialPointsDataFrame}.
}
\description{
Background sample selection.
}
\details{
{First, the function determines the unique pixel coordinates for \emph{x} based on the dimensions of \emph{y} and retrieves
\emph{n}, random background samples where \emph{n} is determined by \emph{nr.samples}. If \emph{sampling.method} is set to \emph{"random"},
the function will return the selected samples as a \emph{SpatialPoints} object. However, if \emph{sampling.method} is set to \emph{"pca"}, the
function performs a Principal Components Analysis (PCA) over \emph{y} to evaluate the similarity between the samples associated to \emph{x} and
the initial set of random samples. To achieve this, the function selects the most important Principal Components (PC's) using the kaiser rule
(i.e. PC's with eigenvalues greater than 1) and, for each PC, estimates the median and the Median Absolute Deviation (MAD) based on the samples
of related ot each unique identifier in \emph{z}). Based on this data, the function selects background samples where the difference between their
variance and the variance of the region samples exceeds the absolute difference between the median and the MAD. Finally, the algorithm filteres out
all the background samples that were not selected by all sample regions. The ouptut is a \emph{SpatialPointsDataFrame} containing the selected samples
and the corresponding \emph{y} values. If \emph{nr.samples} is not provided all background pixels are considered.}
}
\examples{
{

 require(raster)

 # read raster data
 file <- list.files(system.file('extdata', '', package="rsMove"), 'ndvi.tif', full.names=TRUE)
 r.stk <- stack(file)

 # read movement data
 data(shortMove)

 # find sample regions
 label <- labelSample(shortMove, 30, agg.radius=30, nr.pixels=2)

 # select background samples (pca)
 bSamples <- backSample(shortMove, r.stk, label, sampling.method='pca')

 # select background samples (random)
 bSamples <- backSample(shortMove, r.stk, sampling.method='random')

}
}
\seealso{
\code{\link{labelSample}} \code{\link{hotMove}} \code{\link{dataQuery}}
}
