% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_overlap.R
\name{day_overlap}
\alias{day_overlap}
\title{day_overlap}
\usage{
day_overlap(x)
}
\arguments{
\item{x}{Observation time. Object of class \emph{POSIXct}, \emph{POSIXlt}, or \emph{POSIXt}.}
}
\value{
{A \emph{list} containing:
 \itemize{
 \item{\emph{day.cover} - Percent of the day with recorded timestamps.}
 \item{\emph{day.overlap} - Percent of the day covered between the first and last timestamp.}
}
}
}
\description{
Estimate the number of the day covered by timestamps.
}
\details{
{The function estimates how much of a day is covered
by the timestamps in \emph{x}, which correspond to the timestamps
in a movement dataset. Note that \emph{x} is not accepted if
containing data from more than one day.}
}
\examples{
{

# load samples
multiMove <- read.csv(system.file('extdata', 'multiMove.csv', package="rsMove"))

# data-time of species observations
times = strptime(multiMove$timestamp,format="\%Y-\%m-\%d \%H:\%M:\%S")

# subset to first unique day
days = as.Date(times)
times = times[which(days == unique(days)[1])]

# extract regions
day_overlap(date)])

}
}
