% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotMove.R
\name{hotMove}
\alias{hotMove}
\title{hotMove}
\usage{
hotMove(x = x, y = y, pxr = pxr, shp = F, cs = NULL)
}
\arguments{
\item{x}{Vector of x coordinates.}

\item{y}{Vector of y coordinates.}

\item{pxr}{Resolution of the grid used for region labelling. Unir depends on the projection of the input.}

\item{shp}{Logical. If TRUE the function returns a shapefile defined by the convex hull of the samples within each region. Default is FALSE.}

\item{cs}{If shp=TRUE then a valid CRS indicating the projection of the input coordinates is required.}
}
\value{
Returns a vector containing numeric labels that represent unique sample regions. If 'shp' is in use polygons are provided in addition.
}
\description{
Identifies hotspots of samples and labels these accordingly. It build a matrix for a given spatial extent and resolution converting these samples into unique pixel coordinates. Then, the spatial connectivity of these pixels is evaluated using a 8-neighboor connected component labelling algorithm.
}
\examples{
\dontrun{
# reference data
whiteStork <- data(WhiteStork_2013.rdata)

# lat/lon WGS84
outProj <-  CRS("+init=epsg:4978")

# extract regions
hm <- hotMove(x=whiteStork$lon, y=whiteStork=lon, pxr=0.01, shp=T, cs=outProj)

# plot shapefile
plot(hm$polygons)
}
}
\seealso{
\code{\link{sampleMove}}, \code{\link{cRS}}
}
