% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{intime}
\alias{intime}
\title{storeLandsat}
\usage{
intime(x, ij, oj, z)
}
\arguments{
\item{x}{matrix with data to interpolate.}

\item{ij}{Numeric vector with julian days of input data.}

\item{oj}{Numeric vector with julian days of the output data.}

\item{z}{Two-element, numeric vector with temporal buffer.}
}
\value{
A matrix.
}
\description{
Time-adjusted, linear interpolation.
}
\details{
{For each row in x, the function finds the two
nearest, non NA values to the target dates in ij retrieved
before and after this day. Then, the function linearly
interpolates the missing value linearly.}
}
