% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelSample.R
\name{labelSample}
\alias{labelSample}
\title{labelSample}
\usage{
labelSample(x = x, y = y, rad = rad, npx = 2, var = NULL,
  pxr = NULL, layer = NULL)
}
\arguments{
\item{x}{Vector of x coordinates.}

\item{y}{Vector of y coordinates.}

\item{rad}{Minimum radius within which sample regions are aggregated. Unit depends on the projection of the data.}

\item{npx}{Minimum number of pixels wihin a region. Before turning the final sample regions, the function removes regions where the sample count is below the specified value. Removing these samples allows to account for sampling anomalies. The default is 2.}

\item{var}{Optional. Object of class "matrix" or "data.frame". It provides additional numeric variables (e.g. time spent) that wiill be average on a per pixel basis and added to the output.}

\item{pxr}{Pixel resolution used to convert between geographic and pixel coordinates. Required if "layer" is not provided.}

\item{layer}{Raster object which can be provided as an alternative to "pxr".}
}
\value{
Matrix containing the coordinates for selected pixels (x, y) accompanied by their matrix index (index), region ID (region). When provided, the output will contain a summary of "var".
}
\description{
Region labeling of samples based on their spatial connectivity. First, the samples are converted to pixel coordinates and the connectivity between neighboring samples is evaluated. The resulting regions are filtered when required removing small sample groups. Then, the pixels occupied by samples are dilated and a new region labelling is performed. Regions which are within close proximity of each other are thus aggregated. Finnally, for each labeled pixel, the function provides mean sample coordinates and the corresponding pixel index and region ID.
}
\examples{
\dontrun{

}
}
\seealso{
\code{\link{sampleMove}}
}
