% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsMove.R
\docType{data}
\name{longMove}
\alias{longMove}
\title{Example data of animal movements during a migration.}
\format{A SpatialPointsDataFrame}
\usage{
data(longMove)
}
\description{
Movement data for one White Stork collected during it migration between Germany and Spain.
}
\details{
\itemize{
  \item{timestamp}{observation timestamp.}
  \item{long}{longitude.}
  \item{lat}{latitude.}
}
}
\keyword{datasets}
