% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_sites.R
\name{map_sites}
\alias{map_sites}
\title{map_sites}
\usage{
map_sites(x, y, z, resolution, min_size = 1)
}
\arguments{
\item{x}{Object of class \emph{SpatVector}.}

\item{y}{Unique individual identifier for each entry in \emph{x}.}

\item{z}{Timestamps of each element in \emph{x} given as a \emph{POSIXct} object.}

\item{resolution}{Maximum distance between data points to identify clusters.}

\item{min_size}{Minimum number of GPS data points per cluster.}
}
\value{
{A \emph{list} containing:
 \itemize{
 \item{\emph{region.id} - Vector reporting on the region each element in \emph{x} belongs to.}
 \item{\emph{region.polygons} - Polygons for each temporal segment in each \emph{region.id.}
 \item{\emph{region.stats} - Statistics for each temporal segment.}
 }
}}
}
\description{
Detection of geographic regions of samples using a pixel based approach.
}
\details{
{The function provides three outputs. First, it labels each
entry in \emph{x} based on their spatial connectivity. Connections are
based on \emph{resolution}, which defines the maximum distance allowed
among members of a group of data points, below which the elements of
that group are treated as a region and labeled with the same unique,
non-zero, numeric identifier. Regions where the number of data points
is below \emph{min_size} are labelled with a 0. Second, for each region,
the function will return a polygon defined by the convex hull of the
composing data points. 0-labeled regions are excluded, as well as those
where the number of data points is less than 2. Third, for each region,
and for each sequence of days with data points within the target region,
we provide the following information and statistics:
\itemize{
 \item{\emph{region_id} - Region unique identifier.}
 \item{\emph{start.date} - Segment unique identifier.}
 \item{\emph{start.date} - First data date.}
 \item{\emph{end.date} - Last data date.}
 \item{\emph{day_cover_mean} - Mean percentage of the days in the segment with movement data}
 \item{\emph{day_cover_sd} - Standard deviation of the days in the segment with movement data.}
 \item{\emph{day_overlap_mean} - Mean percentage of the day across which movement data were recorded.}
 \item{\emph{day_overlap_sd} - Standard deviation of the percentage of the day across which movement data were recorded.}
 \item{\emph{nr_days} - Number of days with movement data.}
 \item{\emph{nr_individuals} - Number of individuals tracked.}
 \item{\emph{nr_records} - Number of movement data records.}
 \item{\emph{data_frequency_mean} - Mean frequency of GPS entries (in minutes).}
 \item{\emph{data_frequency_sd} - Standard deviation of the frequency of GPS entries (in minutes)}
 }
 Day segments are treated separately because they indicate that a given
 area was persistently occupied by one or more tracked individuals.}
}
\examples{
{

require(terra)

# load samples
multiMove <- read.csv(system.file('extdata', 'multiMove.csv', package="rsMove"))

# convert samples to vector
multiMove = vect(multiMove, geom=c("x","y"), crs="EPSG:4326")[1:10000,]

# species identifier (only one individual, so set to 1)
species_id = multiMove$id

# data-time of species observations
date = strptime(multiMove$timestamp,format="\%Y-\%m-\%d \%H:\%M:\%S", tz="GMT")

# extract regions
hm <- map_sites(multiMove, species_id, date, 0.1)

}
}
