% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveCloud.R
\name{moveCloud}
\alias{moveCloud}
\title{moveCloud}
\usage{
moveCloud(x, y, start, end, interval, data.path)
}
\arguments{
\item{x}{Object of class \emph{SpatVector}.}

\item{y}{Object of class \emph{Date} with observation dates of \emph{y}.}

\item{start}{First data from when to download CCF data.}

\item{end}{Last data from when to download CCF data.}

\item{interval}{Daily interval of data download.}

\item{data.path}{Output path for downloading data on cloud cover.}
}
\value{
{A \emph{list} containing: \itemize{
 \item{\emph{x.stats} - CCF statistics at each entry in \emph{x}}
 \item{\emph{r.stats} - CCF statistics at each unique date between\emph{start} and \emph{end} summarized across the elements of \emph{x}}
 \item{\emph{x.plot} - Plot of the data in \emph{x.stats}}.
 \item{\emph{r.plot} - Plot of the data in \emph{r.stats}}.}}
}
\description{
Extract Cloud Cover Fraction (CCF) data for a set of coordinate pairs.
}
\details{
{The function extracts data on daily Cloud Cover Fractions (CCF)
from NASA's Earth Observation (NEO). For a sequence of dates defined by
\emph{start}, \emph{end}, and \emph{interval}, the function downloads the
correspondent CCF data and stores them in  \emph{data.path}. These data,
which have a global coverage and a spatial resolution of 0.1 degrees, will
only be downloaded if they do not already exist in \emph{data.path}. When
checking for existing data, the function will follow a standard naming
convention, so data acquired independently will likely be missed. After
downloading the needed data, the function will extract the CCF fraction
values for all dates at the coordinates in \emph{x}. This will be used to
calculate the mean and standard deviation of the CCF at each date. In
addition, the function will extract CCF data for each date in \emph{y} that
falls between \emph{start} and \emph{end}.}
}
\examples{
\dontrun{

 require(terra)

 # read movement data
 longMove <- read.csv(system.file('extdata',
 'longMove.csv', package="rsMove"))

 # convert observations to vector
 longMove = vect(longMove, geom=c("long","lat"), crs="EPSG:4326")

 # test function for 30 day buffer
 obs.dates <- as.Date(longMove$timestamp)
 c.cover <- moveCloud(shortMove, obs.dates)

}
}
\references{
\url{https://cneos.jpl.nasa.gov/}
}
\seealso{
\code{\link{sMoveRes}} \code{\link{tMoveRes}}
}
