% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveModel.R
\name{moveModel}
\alias{moveModel}
\title{moveModel}
\usage{
moveModel(samples = samples, features = features, runs = runs,
  method = method, control = control)
}
\arguments{
\item{xyz}{matrix provided by labelSample()}
}
\description{
modeling of
}
\examples{
\dontrun{

}
}
