% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proSat.R
\name{proSat}
\alias{proSat}
\title{proSat}
\usage{
proSat(sensor = sensor, var = NULL, dpath = NULL, ref = NULL)
}
\arguments{
\item{var}{Target variables.}

\item{ref}{Object from which an extent can be derived.}

\item{d.path}{Output data path for downloaded data.}

\item{d.source}{Data source. One of "EarthEnv", "GFC", "GSW", "CCI" or "HSM".}
}
\value{
One or multiple raster objects.
}
\description{
Interface to download ecologically relevant data.
}
\details{
{Downloads data from earthenv.org. To check which variables can be downloaded, 
run the function without specifying \emph{var} and specifying \emph{d.source}. This will 
return a data frame list the existing variables for a given data source. Here, the user 
can refer to the column \emph{"code"} to retrieve the keywords that can be passed to 
the function through \emph{var}. The keywords recognized by \emph{d.source} are:
\itemize{
\item{\emph{"EarthEnv"} - EarthEnv project.}
\item{\emph{"GFC"} - Maryland University Global Forest Change.}
\item{\emph{"GSW"} - JRC Global Surface Water.}
\item{\emph{"CCI"} - ESA CCI Global land cover.}
\item{\emph{"HSM"} - JRC Human Settlement map.}}
If \emph{var} contains \emph{"DEM90"} from {"EarthEv"} or any variable from \emph{"GFC"} 
or \emph{"GSW"}, \emph{ref} is required. This will be used to determine which tiles should 
be downloaded. For details on the specifications of the provided datasets please consult 
the referenced websites.}
}
\examples{
{
 
 # return list of variables
 modis.var <- getEnv(d.source="EarthEnv")
 
}
}
\references{
{\url{http://www.earthenv.org/} 
\url{https://earthenginepartners.appspot.com/science-2013-global-forest/} 
\url{https://global-surface-water.appspot.com/} \url{http://ghsl.jrc.ec.europa.eu/} 
\url{http://maps.elie.ucl.ac.be/CCI/viewer/}}
}
\seealso{
\code{\link{getEnv}}
}
