% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runmean2}
\alias{runmean2}
\title{runmean2}
\usage{
runmean2(x, y)
}
\arguments{
\item{x}{A \emph{Numeric} vector.}

\item{y}{A \emph{numeric} element.}
}
\value{
A vector.
}
\description{
NA-sensitive running mean.
}
\details{
{Applies a running mean over \emph{x} with a window size defined by \emph{y}.
Missing values are ignored and/or filled during the compution.}
}
