% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleMove.R
\name{sampleMove}
\alias{sampleMove}
\title{sampleMove}
\usage{
sampleMove(x, y, time, error, method, layer, tUnit)
}
\arguments{
\item{x}{A Vector of x coordinates}

\item{y}{A Vector of y coordinates}

\item{time}{vVlid "POSIXlt" or "POSIXct" type vector with the time of observation for each record}

\item{method}{Selects the appropriate method to estimate the distance between consecutive points. if 'm' it estimates the ecludian distance. If 'deg' it uses the haversine formula}
}
\value{
Data frame with sample coordinates ('x' and 'y'), total time spent per sample ('time' expressed in minutes) and the total number of observations per sample ('count')
}
\description{
Sampling strategy to identify areas where an animal showed little or no movement based on GPS tracking data.
}
\examples{
\dontrun{

# without reference grid
moveData <- data('whiteStork')
output <- sampleMove(input$x, input$y, input$t, 10, 'm')

# with reference grid
rasterData <- data('grid')
output <- sampleMove(input$x, input$y, input$t, 10, 'm', rasterData)
}
}
