% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsMove.R
\docType{data}
\name{shortMove}
\alias{shortMove}
\title{Example data of animal movements during the nesting period.}
\format{
A SpatialPointsDataFrame
}
\usage{
data(shortMove)
}
\description{
Movement data for one White Stork collected within its nesting site.
}
\details{
\itemize{
  \item{x}{x coordinate.}
  \item{y}{y coordinate.}
  \item{date}{observation date.}
  \item{time}{observation time.}
}
}
\keyword{datasets}
