% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tMoveRes.R
\name{tMoveRes}
\alias{tMoveRes}
\title{tMoveRes}
\usage{
tMoveRes(xy = xy, obs.time = obs.time, t.res = t.res, s.res = s.res)
}
\arguments{
\item{xy}{Object of class \emph{SpatialPoints} or \emph{SpatialPointsDataFrame}.}

\item{obs.time}{Object of class \emph{Date}, \emph{POSIXlt} or \emph{POSIXct} with \emph{xy} observation dates.}

\item{t.res}{Temporal resolution.}

\item{s.res}{Spatial resolution.}
}
\value{
A \emph{list}.
}
\description{
{Tool to support the selection of adequate satellite temporal
resoltuon. Evaluates how the change in temporal resolution changes the
amount of samples/sample regions based on a set of coordinate
pairs and their observation dates.}
}
\details{
{Given a vector of temporal resolutions (\emph{t.res}), the function determines
the number of unique pixels and unique pixel groups after their temporal agggation. The
function returns the corresponding pixel indices per resolution showing which
samples would be grouped (\emph{$indices}). The function returns a data frame (\emph{$stats})
and a plot (\emph{$plot}) with the statistics per temporal resolution.}
}
\examples{
{

 require(raster)

 # reference data
 sprj <- crs("+proj=longlat +ellps=WGS84 +no_defs")
 moveData <- read.csv(system.file('extdata', 'latlon_example.csv', package="rsMove"))
 moveData <- SpatialPointsDataFrame(moveData[,2:3], moveData, proj4string=sprj)

 # test function for 5, 10 20 and 30 m
 obs.date <- as.Date(moveData@data$timestamp)
 a.res <- tMoveRes(xy=moveData, obs.time=obs.date, t.res=c(1,8,16), s.res=0.01)

}
}
\seealso{
\code{\link{sMoveRes}}
}
