% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataQuery.R
\name{dataQuery}
\alias{dataQuery}
\title{dataQuery}
\usage{
dataQuery(x = x, y = y, var = var)
}
\arguments{
\item{x}{Vector of x coordinates.}

\item{y}{Vector of y coordinates.}

\item{var}{Object of class RasterStack or RasterBrick containing environmental layers.}
}
\value{
Matrix of environmental variables for each sample. If a sample does not overlap NA is returned.
}
\description{
Query environmental data for x/y coordinate pairs.
}
\examples{
\dontrun{
}
}
