% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnv.R
\name{getEnv}
\alias{getEnv}
\title{getEnv}
\usage{
getEnv(dpath = NULL, d.source = NULL, var = NULL, ref = NULL)
}
\arguments{
\item{dpath}{Output data path for downloaded data.}

\item{d.source}{Data source. One of "EarthEnv", "GFC", "GSW" or "HSM".}

\item{var}{Target variables.}

\item{ref}{Object from which an extent can be derived.}
}
\value{
One or multiple raster objects.
}
\description{
Interface to download ecologically relevant data.
}
\details{
{Downloads data from earthenv.org. To check which variables can be downloaded, 
run the function without specifying \emph{var} and specifying \emph{d.source}. This will 
return a data frame list the existing variables for a given data source. Here, the user 
can refer to the column \emph{"code"} to retrieve the keywords that can be passed to 
the function through \emph{var}. The keywords recognized by \emph{d.source} are:
\itemize{
\item{\emph{"EarthEnv"} - EarthEnv project.}
\item{\emph{"GFC"} - Maryland University Global Forest Change.}
\item{\emph{"GSW"} - JRC Global Surface Water.}
\item{\emph{"HSM"} - JRC Human Settlement map.}}
If \emph{var} contains \emph{"DEM90"} from {"EarthEv"} or any variable from \emph{"GFC"} 
or \emph{"GSW"}, \emph{ref} is required. This will be used to determine which tiles should 
be downloaded. For details on the specifications of the provided datasets please consult 
the referenced websites.}
}
\examples{
{
 
 # return list of variables
 ee.var <- getEnv(source="earthEnv")
 gfc.var <- getEnv(source="GFC")
 gsw.var <- getEnv(source="GSW")
 hsm.var <- getEnv(source="hsm")
 
}
}
\references{
{\url{http://www.earthenv.org/} 
\url{https://earthenginepartners.appspot.com/science-2013-global-forest/} 
\url{https://global-surface-water.appspot.com/} \url{http://ghsl.jrc.ec.europa.eu/}}
}
\seealso{
\code{\link{sMoveRes}}
}
