% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotMoveStats.R
\name{hotMoveStats}
\alias{hotMoveStats}
\title{hotMoveStats}
\usage{
hotMoveStats(rid = rid, time = NULL, aid = NULL, method = "deg")
}
\arguments{
\item{rid}{Output of hotMove or an object with the same structure (i.e. list containing indices under the keyword $indices and region polygons of the class "SpatialPolygon" under the keyword "polygons")}

\item{time}{Optional. Valid "POSIXlt" or "POSIXct" type vector with the time of observation for each record.}

\item{aid}{Optional. Animal unique id. Use to evaluate the number of individuals per region.}

\item{method}{Method used to estimate polygon area. Required if rid contains a shapefile. If "deg" the function assumes that the used projectn is in degrees. If so, the program projects each polygon to the closest UTM zone so that an area can be estimated. Default is 'deg'.}
}
\value{
Statistics for a sample regions derived with hotMove(). The function provides information on: Total number of samples (tns), numbdf of unique individuals (nui), time of the smallest segment (mnt), average time per segment (avt), time of largest segment (mxt), number of time segments (nts) and area of the polygon (ura). Missing information is reported as NA
}
\description{
Provides statistics for the output of hotMove.
}
\examples{
\dontrun{
# reference data
whiteStork <- data(WhiteStork_2013.rdata)

# lat/lon WGS84
outProj <-  CRS("+init=epsg:4978")

# determine regions
hm <- hotMove(x=whiteStork$lon, y=whiteStork=lon, res=0.01, cs=outProj) 

# extract statistics
hotMoveStats(rid=hm, time=whiteStork$time, aid=whiteStork$id, method='m')
}
}
\seealso{
\code{\link{hotMove}},  \code{\link{CRS}}
}
