% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotMoveStats.R
\name{hotMoveStats}
\alias{hotMoveStats}
\title{hotMoveStats}
\usage{
hotMoveStats(rid = rid, time = NULL, aid = NULL, method = "m")
}
\arguments{
\item{rid}{Output of hotMove or an object with the same structure.}

\item{time}{Optional. Valid "POSIXlt" or "POSIXct" type vector with the time of observation for each record.}

\item{aid}{Optional. Animal unique id. Use to evaluate the number of individuals per region.}

\item{method}{Method used to estimate polygon area. Required if rid contains a shapefile. If "deg" the function assumes that the used projectn is in degrees. If so, the program projects each polygon to the closest UTM zone so that an area can be estimated.}
}
\value{
Statistics for a sample regions derived with hotMove(). The function provides information on: Total number of samples (tns), numbdf of unique individuals (nui), time of the smallest segment (mnt), average time per segment (avt), time of largest segment (mxt) and area of the polygon (ura)
}
\description{
Provides statistics for the output of hotMove.
}
\examples{
\dontrun{
hotMove(x=whiteStork$lon, y=whiteStork=lon, res=0.01, ext)
}
}
\seealso{
\code{\link{hotMove}}
}
