% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelSample.R
\name{labelSample}
\alias{labelSample}
\title{labelSample}
\usage{
labelSample(xy = xy, agg.radius = agg.radius, nr.points = NULL,
  nr.pixels = NULL, pixel.res = pixel.res)
}
\arguments{
\item{xy}{Object of class \emph{SpatialPoints} of \emph{SpatialPointsDataFrame}.}

\item{agg.radius}{Minimum radius for pixel aggregation. Unit depends on the projection of the data.}

\item{nr.points}{Minimum number of samples per pixel.}

\item{nr.pixels}{Minimum number of pixels per region.}

\item{pixel.res}{Pixel resolution os a valid raster layer.}
}
\value{
A \emph{vector} of unique identifiers assigning each point in \emph{xy} to their correspondent pixel region. Filtered observations are returned as \emph{NA}.
}
\description{
{Pixel-based labeling of spatialy connected groups of samples for splitting them between training and validation.}
}
\details{
{First, the samples are converted to pixel coordinates and removes pixels with a corresponding number of points greater
than \emph{nr.points}. Then, if \emph{nr.pixels} is set, the connectivity between neighboring samples is evaluated. Internaly, the
function will label groups of pixels based on their connectivety and regions with a pixel count smaller than the one specified by
\emph{nr.pixels} are excluded. Then, the algorithm aggregates nearby regions using a dilation algorithm within the radius specified
by \emph{agg.radius} and proceeds to relable the pixels covered by samples. Finally, this information is used to label the original
samples provided by \emph{xy} based on their corresponding pixel coordinates. This analysis is based on the spatial extent of \emph{xy}
and a given pixel resolution (\emph{pixel.res}). Alternatively, the user may assign a raster object to \emph{pixel.res}.}
}
\examples{
{

 require(raster)

 # read raster data
 r <- raster(system.file('extdata', 'tcb_1.tif', package="rsMove"))

 # read movement data
 moveData <- read.csv(system.file('extdata', 'konstanz_20130804.csv', package="rsMove"))
 moveData <- SpatialPointsDataFrame(moveData[,1:2], moveData, proj4string=crs(r))

 # derive region labels
 labels <- labelSample(xy=moveData, agg.radius=90, nr.pixels=2, pixel.res=30)

}
}
\seealso{
\code{\link{sampleMove}} \code{\link{hotMove}}
}
