% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveModel.R
\name{moveModel}
\alias{moveModel}
\title{moveModel}
\usage{
moveModel(samples = samples, features = features, method = NULL,
  control = NULL, nruns = 1)
}
\arguments{
\item{samples}{Matrix or data frame containing x and y coordinates and sample region id (0 for background).}

\item{features}{Environmental predictors.}

\item{method}{Classification algorithm (see \url{http://topepo.github.io/caret/index.html}. Default is \emph{rf} (Radom Forest).}

\item{control}{Object derived by \emph{trainControl} (see \code{\link[caret]{trainControl}}). Default used out-of-bag (\emph{oob}) accuracies.}

\item{nruns}{Number of runs. Default is 1.}
}
\value{
A \emph{list}.
}
\description{
Spatially stratified predictive modeling.
}
\details{
{For n iterations, where n is the number of unique sample regions, 
the function uses one sample region for validation while the remaining ones 
are used for training. The background samples are split randomly at each 
iteration. The final accuracy, provided as a F1-score for both presence and 
background sampels, is derived from the total of true and false positives 
(\emph{$f1}). Additionaly, for each run, the function returns a model (\emph{$model}) 
which is trained using all the samples. This output can be passed to modelApply().}
}
\examples{
\dontrun{

 # read example data
 moveData <- shapefile(system.file('extdata', 'konstanz_20130805-20130811.shp', package="rsMove"))

 # extract samples
 ot = strptime(moveData$timestamp, "\%Y-\%m-\%d \%H:\%M:\%S")
 output <- sampleMove(xy=moveData, ot=ot, error=10, method='m')

 # read remote sensing data
 rsStack <- stack(list.files(system.file('extdata', '', package="rsMove"), 'tc.*tif', full.names=T))
 
 # retrieve remote sensing data for samples
 rsQuery <- dataQuery(xy=samples,img=rsStack, rd=T)
 
 # identify unique sample regions
 label <- labelSample(xy=rsQuery, rad=90, npx=4, pxr=30)
 
 # derive model predictions
 out <- moveModel(samples=c(rsQuery@coords,label), features=rsQuery@data)

}
}
\seealso{
\code{\link{sampleMove}} \code{\link{labelSample}} \code{\link{backSample}} \code{\link{modelApply}} \code{\link[caret]{train}}
}
