% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveSeg.R
\name{moveSeg}
\alias{moveSeg}
\title{moveSeg}
\usage{
moveSeg(xy = xy, img = img, type = "cont", ot = NULL, threshold = 0.1,
  r.fun = NULL, s.fun = NULL)
}
\arguments{
\item{xy}{Object of class \emph{SpatialPoints} or \emph{SpatialPointsDataFrame}.}

\item{img}{Object of class \emph{RasterLayer}, \emph{RasterStack} or \emph{RasterBrick}.}

\item{type}{Raster data type. One of \emph{cont} (continues) or \emph{cat} (for categorical).}

\item{ot}{Object of class \emph{Date}, \emph{POSIXlt} or \emph{POSIXct} with \emph{xy} observation dates.}

\item{threshold}{Change threshold.}

\item{r.fun}{Raster summary function. Default is a pca.}

\item{s.fun}{Output summary function. Default is mean.}
}
\value{
A \emph{list}.
}
\description{
Remote sensing based point segmentation
}
\examples{
{
 
 require(rgdal)
 require(raster)
 require(sp)
 
 # read movement data
 moveData <- shapefile(system.file('extdata', 'konstanz_20130804.shp', package="rsMove"))
 
 # read raster data
 r <- raster(system.file('extdata', 'tcb_1.tif', package="rsMove"))
 
 # perform directional sampling
 seg <- moveSeg(xy=moveData, img=r, type="cont", threshold=0.1)
 
}
}
\seealso{
\code{\link{timeDirSample}} \code{\link{dataQuery}}
}
