% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsComposite.R
\name{rsComposite}
\alias{rsComposite}
\title{rsComposite}
\usage{
rsComposite(img, img.dates, obs.dates, comp.method = "closest",
  temporal.buffer = NULL)
}
\arguments{
\item{img}{Object of class \emph{RasterSpack} or \emph{RasterBrick}.}

\item{img.dates}{Object of class \emph{Date} with \emph{img} observation dates.}

\item{obs.dates}{Object of class \emph{Date} with reference dates.}

\item{comp.method}{One of "closest" or "phenological".}

\item{temporal.buffer}{Search buffer (expressed in days). The default is 30.}
}
\value{
A \emph{list}.
}
\description{
{Phenological and date driven Pixel Based Compositing (PBC)
of remote sensing data supported by GPS tracking date information.}
}
\details{
{The function uses a multi-layer raster object to build a composite for
a reference date which corresponds to the median of \emph{obs.dates}. Moreover,
the function determines the Median Absolute Deviation (MAD) of \emph{obs.dates}
which determines the temporal buffer that is used to search for usable images.
As an alternative, \emph{temporal.buffer} can be specified manually and will
be required if \emph{obs.dates} consists of a single value. The user can also
specify how the compositing should be done. \emph{comp.method} can be set to:
#' \itemize{
 \item{\emph{closest} - Uses layer with the closest possible date in relaton to the reference date.}
 \item{\emph{phenological} - Uses the layer with the Day of the Year (DoY) in relation to the reference date.}}
The final output of \emph{rsComposite} is a list consisting of:
#' \itemize{
 \item{\emph{composite} - Final image composite}
 \item{\emph{dates} - Temporal composition of the composite reporting on the julian day}
 \item{\emph{pixel.count} - pixel count of unique values in \emph{dates}. Additionally, it reports on NA values.}
 \item{\emph{target.date} - Reference date used during compositing.}
 \item{\emph{temporal.buffer} - Temporal buffer used during compositing.}}
 If \emph{pheno2} is used, for each pixel, the function wilhin estimate a weighted
 mean of the clear pixels within the temporal buffer. The weights represent the
 inverse time difference between the target and the available dates giver higher
 weights to small differences.}
}
\examples{
\dontrun{

 require(raster)

 # read raster data
 file <- list.files(system.file('extdata', '', package="rsMove"), 'tc.*tif', full.names=TRUE)
 rsStk <- stack(file)
 rsStk <- stack(rsStk, rsStk, rsStk) # dummy files for the example

 # raster dates
 img.dates = seq.Date(as.Date("2013-01-01"), as.Date("2013-12-31"), 45)

 # target date
 obs.dates = as.Date("2013-06-01")

 # build composite
 r.comp <- rsComposite(rsStk, img.dates, obs.dates, comp.method="closest", temporal.buffer=90)

}
}
\seealso{
\code{\link{imgInt}} \code{\link{dataQuery}} \code{\link{proSat}}
}
