% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleMove.R
\name{sampleMove}
\alias{sampleMove}
\title{sampleMove}
\usage{
sampleMove(x = x, y = y, tim = time, error = error, method = method,
  tUnit = tUnit, layer = NULL)
}
\arguments{
\item{x}{Vector of x coordinates}

\item{y}{Vector of y coordinates}

\item{method}{Selects the appropriate method to estimate the distance between consecutive points. if 'm' it estimates the ecludian distance. If 'deg' it uses the haversine formula. Default is 'deg'.}

\item{time}{Valid "POSIXlt" or "POSIXct" type vector with the time of observation for each record}
}
\value{
Data frame with sample coordinates ('x' and 'y'), total time spent per sample ('time' expressed in minutes) and the total number of observations per sample ('count')
}
\description{
Sampling strategy to identify areas where an animal showed little or no movement based on GPS tracking data.
}
\examples{
\dontrun{
# reference data
whiteStork <- data(WhiteStork_2013.rdata)

# sampling without reference grid
output <- sampleMove(whiteStork$lon, whiteStork$lat, whiteStork$time, 7, 'm')

# with reference grid
refGrid<- data('constanceGrid')
output <- sampleMove(whiteStork$lon, whiteStork$lat, whiteStork$time, 7, 'm', refGrid)

# compare original vs new samples
plot(whiteStork$lon, whiteStork$lat, col="black", pch=16)
points(x, y, col="red", pch=15)
}
}
