% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sMoveRes.R
\name{sMoveRes}
\alias{sMoveRes}
\title{sMoveRes}
\usage{
sMoveRes(x, y)
}
\arguments{
\item{x}{Object of class \emph{spatVector}.}

\item{y}{Numeric vector with target  spatial resolutions.}
}
\value{
A \emph{list}.

{The function returns a list with:
\itemize{
 \item{\emph{stats} - Summary statistics reporting on the number
 of unique samples and sample regions per spatial resolution.}
 \item{\emph{plot} - Plot representing the change in number of
 samples and sample regions per spatial resolution.}}}
}
\description{
Analysis of GPS data losses with the choice spatial resolutions.
}
\details{
{The function simulates how many unique combinations of animal
observations and environmental data would be preserved when choosing an
environmental dataset with the resolution given in \emph{y}. It also
estimates simulates how many unique pixel regions (i.e., groups of
spatially connected pixels) would be preserved after accounting for
pseudo-replication. Finally, For each spatial resolution, the function
reports the 'pixel ratio' and the 'region ratio', i.e., the number of
pixels and regions divided by the number of GPS records.}
}
\examples{
{

 require(terra)

 # read movement data
 shortMove = read.csv(system.file('extdata',
 'shortMove.csv', package="rsMove"))

 # convert observations to vector
 shortMove = vect(shortMove, geom=c("x","y"), crs="EPSG:32632")

 # test function for 5, 10 20 and 30 m
 a.res = sMoveRes(shortMove, c(5, 10, 20, 30))

}
}
\seealso{
\code{\link{tMoveRes}}
}
