% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specVar.R
\name{specVar}
\alias{specVar}
\title{specVar}
\usage{
specVar(img = img, xy = NULL, pixel.res = pixel.res)
}
\arguments{
\item{img}{Object of class \emph{RasterLayer}.}

\item{xy}{Object of class \emph{SpatialPoints} or \emph{SpatialPointsDataFrame}.}

\item{pixel.res}{vector of target resolutions.}
}
\value{
A \emph{list}.
}
\description{
{Tool to support the selection of adequate satellite spatial resoltuon. Evaluates
how the spectral variability within a pixel change with the change in spatial resolution.}
}
\details{
{Given a raster object (\emph{img}), the function determines how degrading its spatial resolution impacts
our ability to perceive the complexity of the landscape. For each pixel resolution given by \emph{pixel.res}, The
function resamples \emph{img} and estimates the Mean Absolute Error (MAE) for each pixel based on the differences
between the original and aggregated values. If a point shapefile is provided (\emph{xy}), the function will only
report on the values that overlap with it. The output of the function consists of:
\itemize{
 \item{\emph{mae} - MAE for all pixels or for the points within \emph{xy}.}
 \item{\emph{pixel.optimal} - raster reporting on the resolution with the lowest MAE for each pixel.}
 \item{\emph{pixel.optimal.stats} - Proportion of samples per resolution derived from \emph{pixel.optimal}.}
 \item{\emph{plot} - boxplots of the variability of the MAE per resolution.}
 \item{\emph{sample.optial} - optimal resolution per sample (available if \emph{xy} is provided).}}}
}
\examples{
\dontrun{

 require(raster)

 # read raster data
 r <- raster(system.file('extdata', 'tcb_1.tif', package="rsMove"))

 # read movement data
 moveData <- read.csv(system.file('extdata', 'konstanz_20130804.csv', package="rsMove"))
 moveData <- SpatialPointsDataFrame(moveData[,1:2], moveData, proj4string=crs(r))

 # apply function
 s.var <- specVar(img=r, xy=moveData, pixel.res=60)

}
}
\seealso{
\code{\link{tMoveRes}} \code{\link{sMoveRes}}
}
