% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeDir.R
\name{timeDir}
\alias{timeDir}
\title{timeDir}
\usage{
timeDir(xy = NULL, obs.dates = obs.dates, img = NULL, env.data = NULL,
  env.dates = env.dates, window.size = NULL, sample.direction = NULL,
  stat.fun = NULL, min.count = 2)
}
\arguments{
\item{xy}{Object of class "SpatialPoints" or "SpatialPointsDataFrame".}

\item{obs.dates}{Object of class \emph{Date} with \emph{xy} observation dates.}

\item{img}{Object of class}

\item{env.data}{Object of class \emph{RasterStack} or \emph{RasterBrick} or \emph{data.frame}.}

\item{env.dates}{Object of class \emph{Date} with \emph{env.data} observation dates.}

\item{window.size}{Temporal moving window size (expressed in days).}

\item{sample.direction}{One of \emph{forward}, \emph{backward} or \emph{both}. Default is \emph{both}.}

\item{stat.fun}{Output statistical metric.}

\item{min.count}{Minimum number of samples required by \emph{stat.fun}. Default is 2.}
}
\value{
A \emph{vector} with a requested statistical metric for each point in \emph{xy}.
}
\description{
Analysis of environmental change in time for a set of coordinate pairs.
}
\details{
{This function evaluates how environmental conditions change in time along a movement track.
First, for each point in \emph{xy}, the function compares its observation date (\emph{obs.dates}) against
the acquisition dates (\emph{env.dates}) of \emph{env.data} to select non \emph{NA} timesteps within a
pre-defined temporal window (\emph{window.size}). The user can chose to only consider time steps before
(\emph{backward}) or after (\emph{forward} the target observation time or look at both directios (\emph{both}).
If the latest is chosen, the function applies \emph{window.size} equally to both directions. After selecting
adequate temporal information for each data point, a statistical metric is estimated. The statistical metric
is provided by (\emph{stat.fun}). By default, the slope is reported from a linear regression between the
acquisition times of \emph{env.data} and their corresponding values. When providing a new function, set x
for \emph{env.dates} and y for \emph{env.data}.}
}
\examples{
{

 require(raster)

 # read raster data
 file <- list.files(system.file('extdata', '', package="rsMove"), 'tc.*tif', full.names=TRUE)
 rsStk <- stack(file)
 rsStk <- stack(rsStk, rsStk, rsStk) # dummy files for the example

 # read movement data
 moveData <- read.csv(system.file('extdata', 'konstanz_20130804.csv', package="rsMove"))
 moveData <- SpatialPointsDataFrame(moveData[,1:2], moveData, proj4string=crs(rsStk))

 # raster dates
 r.dates <- seq.Date(as.Date("2013-08-01"), as.Date("2013-08-09"), 1)

 # sample dates
 obs.dates <- as.Date(moveData@data$date)

 # perform directional sampling
 of <- function(x,y) {lm(y~x)$coefficients[2]}
 time.env <- timeDir(xy=moveData, obs.dates=obs.dates, env.data=rsStk,
 env.dates=r.dates, window.size=10, sample.direction="backward", stat.fun=of)

}
}
\seealso{
\code{\link{spaceDir}} \code{\link{dataQuery}} \code{\link{imgInt}}
}
