% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tMoveRes.R
\name{tMoveRes}
\alias{tMoveRes}
\title{tMoveRes}
\usage{
tMoveRes(x, y, time.res, pixel.res)
}
\arguments{
\item{x}{\emph{spatVector}.}

\item{y}{\emph{Date} vector with observation dates or each entry in \emph{x}.}

\item{time.res}{Vector of temporal resolutions (expressed in days).}

\item{pixel.res}{Spatial resolution (unit depends on spatial projection).}
}
\value{
{A \emph{list} containing:
\itemize{
 \item{\emph{stats} - Summary statistics reporting on
 the number of temporal widows, unique samples and unique
 sample regions per temporal resolution.}
 \item{\emph{summary.plot} - Plot representing the change in number
 of unique pixels and pixel regions per temporal resolution.}
 \item{\emph{temporal.plot} - Plot representing the change in number
 of unique pixels and pixel regions per temporal resolution and time step.}}}
}
\description{
Analysis of GPS data losses with the choice temporal resolutions.
}
\details{
{For each spatial resolution given by  \emph{pixel.res}, and for
each temporal resolutions given by \emph{time.res}, the function simulates
the number of unique pixels and pixel regions preserved after accounting
for pseudo-replication at each hypothetical time step, assuming that the
GPS data until the next date with available environmental data is
aggregated into unique pixels. Finally, For each temporal aggregation
window, the function reports the 'pixel ratio' and the 'region ratio',
i.e., the number of pixels and regions divided by the number of GPS records.}
}
\examples{
{

 require(terra)

 #'  # read movement data
 longMove = read.csv(system.file('extdata',
 'longMove.csv', package="rsMove"))

 # convert observations to vector
 longMove = vect(longMove, geom=c("long","lat"), crs="EPSG:4326")

 # test function for intervals of 1, 8 and 16 days (e.g. of MODIS data)
 obs.date = as.Date(longMove$timestamp)
 a.res = tMoveRes(longMove, obs.date, c(1,8,16), 0.1)

}
}
\seealso{
\code{\link{sMoveRes}}
}
