% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tMoveRes.R
\name{tMoveRes}
\alias{tMoveRes}
\title{tMoveRes}
\usage{
tMoveRes(xy = xy, dpath = dpath, p.res = T)
}
\arguments{
\item{xy}{Object of class \emph{SpatialPoints} or \emph{SpatialPointsDataFrame}.}

\item{dpath}{Output data path for downloaded data.}

\item{p.res}{Should the output be ploted on screen? Default is TRUE.}
}
\value{
A \emph{list}.
}
\description{
Provides historical information on cloud cover.
}
\details{
{This function makes use of the global cloud cover statistics provided 
through EarthEnv (\url{http://www.earthenv.org/cloud}) to extract base statistics 
for a shapefile containing GPS tracking data. The function starts by downloading 
monthly means of fractional cloud cover. This requires 9.63 Gb of space on the disk. 
If the data already exists in \emph{dpath} this step will ignored. If the data 
already exists, please assure it has not been modified. Then, the values for each 
point are extracted (\emph{$point.data}) and used to derive the min, max, mean and 
sd for each month (\emph{$stats}). This statistics are then used to build a plot 
with ggplot expressing the mean and maximum cloud cover fraction per month. This 
plot will be added to the output (\emph{$plot}). If \emph{p.res} is TRUE, it is 
also ploted on screen.}
}
\examples{
\dontrun{
 
 require(raster)
 
 # read raster data
 r <- raster(system.file('extdata', 'tcb_1.tif', package="rsMove"))
 
 # read movement data
 moveData <- read.csv(system.file('extdata', 'konstanz_20130804.csv', package="rsMove"))
 moveData <- SpatialPointsDataFrame(moveData[,1:2], moveData, proj4string=crs(r))
 
 # test function for 5, 10 20 and 30 m
 a.res <- tMoveRes(xy=moveData, dpath='.')
 
}
}
\references{
\url{https://doi.org/10.1371/journal.pbio.1002415}
}
\seealso{
\code{\link{sMoveRes}}
}
